requirements.txt
google-generativeai>=0.3.0
spotipy>=2.22.0
requests>=2.31.0
python-dotenv>=1.0.0
icalendar>=5.0.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
